Add files via upload
The first step in the analysis is to filter the reads by quality, you should find the optimal parameters through exploring filtering your data with different parameters. Two options are available for this step: When the reads are paired-end use first cutadapt.h to remove the primers, then flash.sh to merge the pairs and finally filtering_SE.R. If the reads are single-end do the same but skipping the flash.sh step. After having the reads filtered the second step is to infer sequences variants and perform taxonomic assignment using DADA2 and SILVA database. To do this use the Inf_Chim_tax.R script. However, if your data make the analysis run out of memory use the InfChim_tax_2.R script instead.