Add files via upload
This file contains the instructions to merge the results from the different studies analyzed, filter the samples and taxa according to rarefaction curves information, and perform the alpha and beta diversity inference and statistical analysis of the samples using Shannon, Simpson, Chao1 and  PCoA coupled with Weighted Unifrac distance. This is not a script to run automatically because you have to be aware of the output of your analysis in each step to decide the parameters of the next one. This is just a guide of the steps in the analysis and you should run it line by line. It also contains the codes to make beautiful graphics of the results using ggplot2.