Add files via upload
Nex thing you can do after analyzing changes in the global diversity of your samples is to focus in specific taxons that may be associated to the disease or treatment you are studying. That could be assessed with a differential abundance analysis. This file "DifferentialAbundance.R" contains a set of instructions to perform the differential abundance analysis using DESeq2. Like the previous file, this is not a script to run automatically, you should run it line by line and review your results in each step.